finalizeConfiguration: check invariants on converted ACLs
We have had a number of issues (see batfish/batfish#8388 for example)
where an ACL checks a protocol-specific property but does not actually
restrict to packets of the correct protocol. This can lead to strange
behavior, such as allowing flows that have ICMP Type 128 but are not
ICMP packets - mainly in symbolic tools like reachability.

Check converted ACLs when assertions are enabled. Note that we only check ACLs,
rather than individual MatchExprs -- because the latter can be composed and
it's difficult to tell when we're at an abstraction boundary when these
invariants should be enforced.

ASA's object groups are not that well formed, so skip this validation
on ASA for now.

commit-id:aca59b81