rescan hash bucket when it changes
two threads may race to create the same node. after we identify the hash bucket, we scan to see if the node already exists. if not,
we try to insert. when we do insert, if the hash bucket has changed, another thread may have inserted a copy of the node, so we have to rescan.